+++ pages/api/plug/plug-all-off.ts
@@ -2,21 +2,21 @@ import { turnOff } from "tp-link-tapo-connect";
 import { getAllPlugs } from "pages/api/plug/statuses";
 
 export default async function handler(req, res) {
-	try {
-		const plugs = await getAllPlugs();
+  try {
+    const plugs = await getAllPlugs();
 
-		Object.keys(plugs).forEach(async (plug) => {
-			const token = plugs[plug].token;
-			await turnOff(token);
-		});
+    Object.keys(plugs).forEach(async (plug) => {
+      const token = plugs[plug].token;
+      await turnOff(token);
+    });
 
-		const statusesOut = await getAllPlugs();
+    const statusesOut = await getAllPlugs();
 
-		res.status(200).json({
-			statusesOut,
-		});
-	} catch (e) {
-		console.error(e);
-		res.status(400).json({ message: "Error", error });
-	}
+    res.status(200).json({
+      statusesOut,
+    });
+  } catch (e) {
+    console.error(e);
+    res.status(400).json({ message: "Error", e });
+  }
 }
