+++ pages/dashboard/index.tsx
@@ -8,164 +8,163 @@ import DefaultLayout from "layouts/default";
 import useInterval from "hooks/useInterval";
 
 import {
-	getTimeValues,
-	getDayFromDayNumber,
-	getSeasonFromYear,
-	getMonthNameFromMonthNumber,
+  getTimeValues,
+  getDayFromDayNumber,
+  getMonthNameFromMonthNumber,
 } from "utils/time";
 
-const Clock = ({ minutes, hours }) => {
-	const isPm = hours > 12;
+import getSeasonFromYear from "utils/getSeasonFromYear";
 
-	return (
-		<div>
-			{hours > 12 ? hours - 12 : hours}:{String(minutes).padStart(2, "0")}{" "}
-			{isPm ? "pm" : "am"}
-		</div>
-	);
+const Clock = ({ minutes, hours }) => {
+  const isPm = hours > 12;
+
+  return (
+    <div>
+      {hours > 12 ? hours - 12 : hours}:{String(minutes).padStart(2, "0")}{" "}
+      {isPm ? "pm" : "am"}
+    </div>
+  );
 };
 
 const Temperature = ({ temperature, type = "in" }) => {
-	return (
-		<div>
-			<div>Temp. {type}:</div>
-			{temperature !== undefined && (
-				<Fragment>
-					{temperature.toFixed(1)}
-					<sup>&deg;</sup>C
-				</Fragment>
-			)}
-		</div>
-	);
+  return (
+    <div>
+      <div>Temp. {type}:</div>
+      {temperature !== undefined && (
+        <Fragment>
+          {temperature.toFixed(1)}
+          <sup>&deg;</sup>C
+        </Fragment>
+      )}
+    </div>
+  );
 };
 
 const DateOfMonth = ({ date }) => {
-	return <div>{date}</div>;
+  return <div>{date}</div>;
 };
 
 const Month = ({ month }) => {
-	const monthName = getMonthNameFromMonthNumber(month);
-	return <div>{monthName}</div>;
+  const monthName = getMonthNameFromMonthNumber(month);
+  return <div>{monthName}</div>;
 };
 
 const Year = ({ year }) => {
-	return <div>{year}</div>;
+  return <div>{year}</div>;
 };
 
 const DayOfWeek = ({ day }) => {
-	return <div>{getDayFromDayNumber(day)}</div>;
+  return <div>{getDayFromDayNumber(day)}</div>;
 };
 
 const Season = ({ season }) => {
-	return <div>{season}</div>;
+  return <div>{season}</div>;
 };
 
 export default function Dashboard() {
-	const [hours, setHours] = useState(0);
-	const [minutes, setMinutes] = useState(0);
-	const [year, setYear] = useState(0);
-	const [day, setDay] = useState(0);
-	const [month, setMonth] = useState(0);
-	const [season, setSeason] = useState(0);
-	const [date, setDate] = useState(0);
-	const [temperatureInside, setTemperatureInside] = useState(0);
-	const [temperatureOutside, setTemperatureOutside] = useState(0);
+  const [hours, setHours] = useState(0);
+  const [minutes, setMinutes] = useState(0);
+  const [year, setYear] = useState(0);
+  const [day, setDay] = useState(0);
+  const [month, setMonth] = useState(0);
+  const [season, setSeason] = useState("");
+  const [date, setDate] = useState(0);
+  const [temperatureInside, setTemperatureInside] = useState(0);
+  const [temperatureOutside, setTemperatureOutside] = useState(0);
 
-	useEffect(() => {
-		const {
-			currentHour,
-			currentMinute,
-			currentYear,
-			currentDay,
-			currentMonth,
-			currentDate,
-		} = getTimeValues();
+  useEffect(() => {
+    const {
+      currentHour,
+      currentMinute,
+      currentYear,
+      currentDay,
+      currentMonth,
+      currentDate,
+    } = getTimeValues();
 
-		const currentSeason = getSeasonFromYear(currentYear);
+    const currentSeason = getSeasonFromYear(currentYear);
 
-		setHours(currentHour);
-		setMinutes(currentMinute);
-		setYear(currentYear);
-		setDay(currentDay);
-		setMonth(currentMonth);
-		setSeason(currentSeason);
-		setDate(currentDate);
-	}, []);
+    setHours(currentHour);
+    setMinutes(currentMinute);
+    setYear(currentYear);
+    setDay(currentDay);
+    setMonth(currentMonth);
+    setSeason(currentSeason);
+    setDate(currentDate);
+  }, []);
 
-	const getOutsideTemperature = async () => {
-		try {
-			const result = await fetch("/api/weather/get-weather")
-				.then((response) => response.json())
-				.then((json) => json.result);
+  const getOutsideTemperature = async () => {
+    try {
+      const result = await fetch("/api/weather/get-weather")
+        .then((response) => response.json())
+        .then((json) => json.result);
 
-			setTemperatureOutside(result.temp.value);
-		} catch (e) {
-			console.error(e);
-		}
-	};
+      setTemperatureOutside(result.temp.value);
+    } catch (e) {
+      console.error(e);
+    }
+  };
 
-	const getInsideTemperature = async () => {
-		try {
-			const { temperature } = await fetch(
-				"/api/weather/get-pi-sensor"
-			).then((response) => response.json());
+  const getInsideTemperature = async () => {
+    try {
+      const { temperature } = await fetch(
+        "/api/weather/get-pi-sensor"
+      ).then((response) => response.json());
 
-			setTemperatureInside(temperature);
-		} catch (e) {
-			console.error("error in getInsideTemperature", e);
-			const { sensor_data } = await fetch(
-				"/api/hasura/get-weather?number=1"
-			).then((response) => response.json());
+      setTemperatureInside(temperature);
+    } catch (e) {
+      console.error("error in getInsideTemperature", e);
+      const { sensor_data } = await fetch("/api/hasura/get-weather?number=1")
+        .then((response) => response.json())
+        .catch((e) => console.error("error in getting weather from api"));
+      console.log(sensor_data);
+      setTemperatureInside(sensor_data[0]?.temperature);
+    }
+  };
 
-			setTemperatureInside(sensor_data[0].temperature);
-		}
-	};
+  useEffect(() => {
+    const getOutsideTemp = async () => await getOutsideTemperature();
+    getOutsideTemp();
+  }, []);
 
-	useEffect(async () => {
-		await getOutsideTemperature();
-	}, []);
+  useEffect(() => {
+    const getInsideTemp = async () => await getInsideTemperature();
+    getInsideTemp();
+  }, []);
 
-	useEffect(async () => {
-		await getInsideTemperature();
-	}, []);
+  useInterval(() => {
+    const { currentHour, currentMinute } = getTimeValues();
+    setHours(currentHour);
+    setMinutes(currentMinute);
+  }, 1000);
 
-	useInterval(() => {
-		const { currentHour, currentMinute } = getTimeValues();
-		setHours(currentHour);
-		setMinutes(currentMinute);
-	}, 1000);
+  useInterval(async () => {
+    const { currentYear, currentDay, currentMonth } = getTimeValues();
+    setMonth(currentMonth);
+    setDay(currentDay);
+    setYear(currentYear);
+  }, 3000);
 
-	useInterval(async () => {
-		const { currentYear, currentDay, currentMonth } = getTimeValues();
-		setMonth(currentMonth);
-		setDay(currentDay);
-		setYear(currentYear);
-	}, 3000);
+  // Every 3 minutes, call Weather API to get the outside weather.
+  useInterval(async () => {
+    // GET THE TEMPERATURE FROM THE API
+    await getOutsideTemperature();
 
-	// Every 3 minutes, call Weather API to get the outside weather.
-	useInterval(async () => {
-		// GET THE TEMPERATURE FROM THE API
-		await getOutsideTemperature();
+    await getInsideTemperature();
+  }, 1000 * 3 * 60);
 
-		await getInsideTemperature();
-	}, 1000 * 3 * 60);
-
-	return (
-		<DefaultLayout>
-			<Head>
-				<title>Dashboard</title>
-			</Head>
-
-			<div className={styles.container}>
-				<DayOfWeek day={day} />
-				<DateOfMonth date={date} />
-				<Month month={month} />
-				<Year year={year} />
-				<Season season={season} />
-				<Clock hours={hours} minutes={minutes} />
-				<Temperature temperature={temperatureInside} />
-				<Temperature temperature={temperatureOutside} type="out" />
-			</div>
-		</DefaultLayout>
-	);
+  return (
+    <DefaultLayout title="">
+      <div className={styles.container}>
+        <DayOfWeek day={day} />
+        <DateOfMonth date={date} />
+        <Month month={month} />
+        <Year year={year} />
+        <Season season={season} />
+        <Clock hours={hours} minutes={minutes} />
+        <Temperature temperature={temperatureInside} />
+        <Temperature temperature={temperatureOutside} type="out" />
+      </div>
+    </DefaultLayout>
+  );
 }
